rework inheritance
