+++ packages/alsatian/core/matchers/index.ts
@@ -1,3 +1,4 @@
+import { BufferMatcher } from "./buffer-matcher";
 import { ContainerMatcher } from "./container-matcher";
 import { EmptyMatcher } from "./empty-matcher";
 import { FunctionMatcher } from "./function-matcher";
@@ -10,6 +11,7 @@ import { SpyCallCountType } from "./spy-call-count-type";
 import { StringMatcher } from "./string-matcher";
 
 export {
+	BufferMatcher,
 	ContainerMatcher,
 	EmptyMatcher,
 	FunctionMatcher,
